+++ app/components/navigation.top.main.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { useNavigate } from "react-router";
 
 import { authClient } from "~/lib/auth.client";
-import queryClient from "~/lib/query/query.client";
+import queryClient from "~/lib/query.client";
 
 import { cn } from "~/lib/utils";
