+++ app/routes/post/components/comments.section.tsx
@@ -11,7 +11,7 @@ import { z } from "zod";
 import { cn } from "~/lib/utils";
 
 import axiosInstance from "~/lib/axios.instance";
-import queryClient from "~/lib/query/query.client";
+import queryClient from "~/lib/query.client";
 
 import type { Session } from "~/lib/auth.client";
 import type { Post, User } from "~/lib/types/shared";
