+++ app/routes/post/page.$post.tsx
@@ -6,7 +6,7 @@ import React from "react";
 import { useLocation, useNavigate } from "react-router";
 
 import axiosInstance from "~/lib/axios.instance";
-import queryClient from "~/lib/query/query.client";
+import queryClient from "~/lib/query.client";
 
 import { authClient } from "~/lib/auth.client";
