+++ app/routes/users/components/posts.card.tsx
@@ -18,7 +18,7 @@ import { authClient } from "~/lib/auth.client";
 import { cn, formatTime } from "~/lib/utils";
 
 import axiosInstance from "~/lib/axios.instance";
-import queryClient from "~/lib/query/query.client";
+import queryClient from "~/lib/query.client";
 
 import type { Post, User } from "~/lib/types/shared";
